The following changes were made:
1) Changed appropriate counter variables to type long int and updated examples

2) Fixed constraint bug:

   2a) Now only elements actually violating a constraint are considered when calculating the value of alpha (the value by which the step must be scaled)
   2b) The value of alpha is now stored in the variable kinmem->stepmul

3) KINSOL now uses inequality constraints:

   constraints(i) == +2.0 -> uu(i) > 0.0
   constraints(i) == -2.0 -> uu(i) < 0.0
   constraints(i) == +1.0 -> uu(i) >= 0.0
   constraints(i) == -1.0 -> uu(i) <= 0.0
   constraints(i) == 0.0 -> NO constraint

4) Fixed bugs in KINInexactNewton and KINLineSearch which caused the step vector to be incorrectly scaled when a constraint was violated - alpha was being scaled

5) Fixed error in KINLineSearch related to beta condition

Note: There will probably be at least one additional update soon.

+++ kinsol/fcmix/fkinbbd.c
@@ -83,7 +83,7 @@ void FKINgcomm(long int Nloc, N_Vector uu, void *f_data)
 
 /* C function FKINBBDOPT to access optional outputs from KBBD_Data */
 
-void FKIN_BBDOPT(long int *lenrpw, long int *lenipw, int *nge)
+void FKIN_BBDOPT(long int *lenrpw, long int *lenipw, long int *nge)
 {
   KBBDPrecGetIntWorkSpace(KBBD_Data, lenipw);
   KBBDPrecGetRealWorkSpace(KBBD_Data, lenrpw);
